copenet_real/src/copenet_real/requirements.txt
absl-py==0.9.0
apptools==4.5.0
backcall==0.1.0
bleach==3.1.0
Bottleneck==1.3.1
cachetools==4.0.0
catkin-pkg==0.4.16
chumpy==0.69
configer==1.3.1
configparser==4.0.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
entrypoints==0.3
envisage==4.9.0
freetype-py==2.1.0.post1
future==0.18.2
glob3==0.0.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
h5py==2.10.0
imageio==2.6.1
imgaug==0.4.0
importlib-metadata==1.5.0
ipdb==0.12.3
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mayavi==4.7.1
meshcat==0.0.18
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.2
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
open3d-python==0.7.0.0
opencv-python==4.2.0.32
pandocfilters==1.4.2
parso==0.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
plotly==4.5.4
plyfile==0.7.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.2
psutil==5.7.0
pyface==6.1.2
pyglet==1.4.0b1
Pygments==2.5.2
PyOpenGL==3.1.5
pyparsing==2.4.6
pyrender==0.1.34
pyrsistent==0.15.7
pytorch-lightning==0.8.5
PyYAML==5.3
pyzmq==19.0.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rospkg==1.2.3
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
tensorboard==2.1.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchgeometry==0.1.2
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
traits==5.2.0
traitsui==6.1.3
trimesh==3.5.16
u-msgpack-python==2.6.0
vtk==8.1.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
zipp==2.1.0
